Compute the Cell Centers of a Cell Mapping Grid

Computes the cell centers and the corresponding cell IDs of a cell mapping grid.

computeGridCenters(lower, upper, blocks)

Arguments

lower
[numeric or integer] The lower limits per dimension.

upper
[numeric or integer] The upper limits per dimension.

blocks
[integer] The number of blocks per dimension.

Value

[data.frame]. A data.frame, which includes the coordinates of the cell centers, as well as the corresponding cell ID (cell.ID).

Examples

computeGridCenters(lower = -10, upper = 10, blocks = c(10, 5, 8))

#>     x1 x2    x3 cell.ID
#> 1   -9 -8 -8.75       1
#> 2   -7 -8 -8.75       2
#> 3   -5 -8 -8.75       3
#> 4   -3 -8 -8.75       4
#> 5   -1 -8 -8.75       5
#> 6    1 -8 -8.75       6
#> 7    3 -8 -8.75       7
#> 8    5 -8 -8.75       8
#> 9    7 -8 -8.75       9
#> 10   9 -8 -8.75      10
#> 11  -9 -4 -8.75      11
#> 12  -7 -4 -8.75      12
#> 13  -5 -4 -8.75      13
#> 14  -3 -4 -8.75      14
#> 15  -1 -4 -8.75      15
#> 16   1 -4 -8.75      16
#> 17   3 -4 -8.75      17
#> 18   5 -4 -8.75      18
#> 19   7 -4 -8.75      19
#> 20   9 -4 -8.75      20
#> 21  -9  0 -8.75      21
#> 22  -7  0 -8.75      22
#> 23  -5  0 -8.75      23
#> 24  -3  0 -8.75      24
#> 25  -1  0 -8.75      25
#> 26   1  0 -8.75      26
#> 27   3  0 -8.75      27
#> 28   5  0 -8.75      28
#> 29   7  0 -8.75      29
#> 30   9  0 -8.75      30
#> 31  -9  4 -8.75      31
#> 32  -7  4 -8.75      32
#> 33  -5  4 -8.75      33
#> 34  -3  4 -8.75      34
#> 35  -1  4 -8.75      35
#> 36   1  4 -8.75      36
#> 37   3  4 -8.75      37
#> 38   5  4 -8.75      38
#> 39   7  4 -8.75      39
#> 40   9  4 -8.75      40
#> 41  -9  8 -8.75      41
#> 42  -7  8 -8.75      42
#> 43  -5  8 -8.75      43
#> 44  -3  8 -8.75      44
#> 45  -1  8 -8.75      45
#> 46   1  8 -8.75      46
#> 47   3  8 -8.75      47
#> 48   5  8 -8.75      48
#> 49   7  8 -8.75      49
#> 50   9  8 -8.75      50
#> 51  -9 -8 -6.25      51
#> 52  -7 -8 -6.25      52
#> 53  -5 -8 -6.25      53
#> 54  -3 -8 -6.25      54
#> 55  -1 -8 -6.25      55
#> 56   1 -8 -6.25      56
#> 57   3 -8 -6.25      57
#> 58   5 -8 -6.25      58
#> 59   7 -8 -6.25      59
#> 60   9 -8 -6.25      60
#> 61  -9 -4 -6.25      61
#> 62  -7 -4 -6.25      62
#> 63  -5 -4 -6.25      63
#> 64  -3 -4 -6.25      64
#> 65  -1 -4 -6.25      65
#> 66   1 -4 -6.25      66
#> 67   3 -4 -6.25      67
#> 68   5 -4 -6.25      68
#> 69   7 -4 -6.25      69
#> 70   9 -4 -6.25      70
#> 71  -9  0 -6.25      71
#> 72  -7  0 -6.25      72
#> 73  -5  0 -6.25      73
#> 74  -3  0 -6.25      74
#> 75  -1  0 -6.25      75
#> 76   1  0 -6.25      76
#> 77   3  0 -6.25      77
#> 78   5  0 -6.25      78
#> 79   7  0 -6.25      79
#> 80   9  0 -6.25      80
#> 81  -9  4 -6.25      81
#> 82  -7  4 -6.25      82
#> 83  -5  4 -6.25      83
#> 84  -3  4 -6.25      84
#> 85  -1  4 -6.25      85
#> 86   1  4 -6.25      86
#> 87   3  4 -6.25      87
#> 88   5  4 -6.25      88
#> 89   7  4 -6.25      89
#> 90   9  4 -6.25      90
#> 91  -9  8 -6.25      91
#> 92  -7  8 -6.25      92
#> 93  -5  8 -6.25      93
#> 94  -3  8 -6.25      94
#> 95  -1  8 -6.25      95
#> 96   1  8 -6.25      96
#> 97   3  8 -6.25      97
#> 98   5  8 -6.25      98
#> 99   7  8 -6.25      99
#> 100  9  8 -6.25     100
#> 101 -9 -8 -3.75     101
#> 102 -7 -8 -3.75     102
#> 103 -5 -8 -3.75     103
#> 104 -3 -8 -3.75     104
#> 105 -1 -8 -3.75     105
#> 106  1 -8 -3.75     106
#> 107  3 -8 -3.75     107
#> 108  5 -8 -3.75     108
#> 109  7 -8 -3.75     109
#> 110  9 -8 -3.75     110
#> 111 -9 -4 -3.75     111
#> 112 -7 -4 -3.75     112
#> 113 -5 -4 -3.75     113
#> 114 -3 -4 -3.75     114
#> 115 -1 -4 -3.75     115
#> 116  1 -4 -3.75     116
#> 117  3 -4 -3.75     117
#> 118  5 -4 -3.75     118
#> 119  7 -4 -3.75     119
#> 120  9 -4 -3.75     120
#> 121 -9  0 -3.75     121
#> 122 -7  0 -3.75     122
#> 123 -5  0 -3.75     123
#> 124 -3  0 -3.75     124
#> 125 -1  0 -3.75     125
#> 126  1  0 -3.75     126
#> 127  3  0 -3.75     127
#> 128  5  0 -3.75     128
#> 129  7  0 -3.75     129
#> 130  9  0 -3.75     130
#> 131 -9  4 -3.75     131
#> 132 -7  4 -3.75     132
#> 133 -5  4 -3.75     133
#> 134 -3  4 -3.75     134
#> 135 -1  4 -3.75     135
#> 136  1  4 -3.75     136
#> 137  3  4 -3.75     137
#> 138  5  4 -3.75     138
#> 139  7  4 -3.75     139
#> 140  9  4 -3.75     140
#> 141 -9  8 -3.75     141
#> 142 -7  8 -3.75     142
#> 143 -5  8 -3.75     143
#> 144 -3  8 -3.75     144
#> 145 -1  8 -3.75     145
#> 146  1  8 -3.75     146
#> 147  3  8 -3.75     147
#> 148  5  8 -3.75     148
#> 149  7  8 -3.75     149
#> 150  9  8 -3.75     150
#> 151 -9 -8 -1.25     151
#> 152 -7 -8 -1.25     152
#> 153 -5 -8 -1.25     153
#> 154 -3 -8 -1.25     154
#> 155 -1 -8 -1.25     155
#> 156  1 -8 -1.25     156
#> 157  3 -8 -1.25     157
#> 158  5 -8 -1.25     158
#> 159  7 -8 -1.25     159
#> 160  9 -8 -1.25     160
#> 161 -9 -4 -1.25     161
#> 162 -7 -4 -1.25     162
#> 163 -5 -4 -1.25     163
#> 164 -3 -4 -1.25     164
#> 165 -1 -4 -1.25     165
#> 166  1 -4 -1.25     166
#> 167  3 -4 -1.25     167
#> 168  5 -4 -1.25     168
#> 169  7 -4 -1.25     169
#> 170  9 -4 -1.25     170
#> 171 -9  0 -1.25     171
#> 172 -7  0 -1.25     172
#> 173 -5  0 -1.25     173
#> 174 -3  0 -1.25     174
#> 175 -1  0 -1.25     175
#> 176  1  0 -1.25     176
#> 177  3  0 -1.25     177
#> 178  5  0 -1.25     178
#> 179  7  0 -1.25     179
#> 180  9  0 -1.25     180
#> 181 -9  4 -1.25     181
#> 182 -7  4 -1.25     182
#> 183 -5  4 -1.25     183
#> 184 -3  4 -1.25     184
#> 185 -1  4 -1.25     185
#> 186  1  4 -1.25     186
#> 187  3  4 -1.25     187
#> 188  5  4 -1.25     188
#> 189  7  4 -1.25     189
#> 190  9  4 -1.25     190
#> 191 -9  8 -1.25     191
#> 192 -7  8 -1.25     192
#> 193 -5  8 -1.25     193
#> 194 -3  8 -1.25     194
#> 195 -1  8 -1.25     195
#> 196  1  8 -1.25     196
#> 197  3  8 -1.25     197
#> 198  5  8 -1.25     198
#> 199  7  8 -1.25     199
#> 200  9  8 -1.25     200
#> 201 -9 -8  1.25     201
#> 202 -7 -8  1.25     202
#> 203 -5 -8  1.25     203
#> 204 -3 -8  1.25     204
#> 205 -1 -8  1.25     205
#> 206  1 -8  1.25     206
#> 207  3 -8  1.25     207
#> 208  5 -8  1.25     208
#> 209  7 -8  1.25     209
#> 210  9 -8  1.25     210
#> 211 -9 -4  1.25     211
#> 212 -7 -4  1.25     212
#> 213 -5 -4  1.25     213
#> 214 -3 -4  1.25     214
#> 215 -1 -4  1.25     215
#> 216  1 -4  1.25     216
#> 217  3 -4  1.25     217
#> 218  5 -4  1.25     218
#> 219  7 -4  1.25     219
#> 220  9 -4  1.25     220
#> 221 -9  0  1.25     221
#> 222 -7  0  1.25     222
#> 223 -5  0  1.25     223
#> 224 -3  0  1.25     224
#> 225 -1  0  1.25     225
#> 226  1  0  1.25     226
#> 227  3  0  1.25     227
#> 228  5  0  1.25     228
#> 229  7  0  1.25     229
#> 230  9  0  1.25     230
#> 231 -9  4  1.25     231
#> 232 -7  4  1.25     232
#> 233 -5  4  1.25     233
#> 234 -3  4  1.25     234
#> 235 -1  4  1.25     235
#> 236  1  4  1.25     236
#> 237  3  4  1.25     237
#> 238  5  4  1.25     238
#> 239  7  4  1.25     239
#> 240  9  4  1.25     240
#> 241 -9  8  1.25     241
#> 242 -7  8  1.25     242
#> 243 -5  8  1.25     243
#> 244 -3  8  1.25     244
#> 245 -1  8  1.25     245
#> 246  1  8  1.25     246
#> 247  3  8  1.25     247
#> 248  5  8  1.25     248
#> 249  7  8  1.25     249
#> 250  9  8  1.25     250
#> 251 -9 -8  3.75     251
#> 252 -7 -8  3.75     252
#> 253 -5 -8  3.75     253
#> 254 -3 -8  3.75     254
#> 255 -1 -8  3.75     255
#> 256  1 -8  3.75     256
#> 257  3 -8  3.75     257
#> 258  5 -8  3.75     258
#> 259  7 -8  3.75     259
#> 260  9 -8  3.75     260
#> 261 -9 -4  3.75     261
#> 262 -7 -4  3.75     262
#> 263 -5 -4  3.75     263
#> 264 -3 -4  3.75     264
#> 265 -1 -4  3.75     265
#> 266  1 -4  3.75     266
#> 267  3 -4  3.75     267
#> 268  5 -4  3.75     268
#> 269  7 -4  3.75     269
#> 270  9 -4  3.75     270
#> 271 -9  0  3.75     271
#> 272 -7  0  3.75     272
#> 273 -5  0  3.75     273
#> 274 -3  0  3.75     274
#> 275 -1  0  3.75     275
#> 276  1  0  3.75     276
#> 277  3  0  3.75     277
#> 278  5  0  3.75     278
#> 279  7  0  3.75     279
#> 280  9  0  3.75     280
#> 281 -9  4  3.75     281
#> 282 -7  4  3.75     282
#> 283 -5  4  3.75     283
#> 284 -3  4  3.75     284
#> 285 -1  4  3.75     285
#> 286  1  4  3.75     286
#> 287  3  4  3.75     287
#> 288  5  4  3.75     288
#> 289  7  4  3.75     289
#> 290  9  4  3.75     290
#> 291 -9  8  3.75     291
#> 292 -7  8  3.75     292
#> 293 -5  8  3.75     293
#> 294 -3  8  3.75     294
#> 295 -1  8  3.75     295
#> 296  1  8  3.75     296
#> 297  3  8  3.75     297
#> 298  5  8  3.75     298
#> 299  7  8  3.75     299
#> 300  9  8  3.75     300
#> 301 -9 -8  6.25     301
#> 302 -7 -8  6.25     302
#> 303 -5 -8  6.25     303
#> 304 -3 -8  6.25     304
#> 305 -1 -8  6.25     305
#> 306  1 -8  6.25     306
#> 307  3 -8  6.25     307
#> 308  5 -8  6.25     308
#> 309  7 -8  6.25     309
#> 310  9 -8  6.25     310
#> 311 -9 -4  6.25     311
#> 312 -7 -4  6.25     312
#> 313 -5 -4  6.25     313
#> 314 -3 -4  6.25     314
#> 315 -1 -4  6.25     315
#> 316  1 -4  6.25     316
#> 317  3 -4  6.25     317
#> 318  5 -4  6.25     318
#> 319  7 -4  6.25     319
#> 320  9 -4  6.25     320
#> 321 -9  0  6.25     321
#> 322 -7  0  6.25     322
#> 323 -5  0  6.25     323
#> 324 -3  0  6.25     324
#> 325 -1  0  6.25     325
#> 326  1  0  6.25     326
#> 327  3  0  6.25     327
#> 328  5  0  6.25     328
#> 329  7  0  6.25     329
#> 330  9  0  6.25     330
#> 331 -9  4  6.25     331
#> 332 -7  4  6.25     332
#> 333 -5  4  6.25     333
#> 334 -3  4  6.25     334
#> 335 -1  4  6.25     335
#> 336  1  4  6.25     336
#> 337  3  4  6.25     337
#> 338  5  4  6.25     338
#> 339  7  4  6.25     339
#> 340  9  4  6.25     340
#> 341 -9  8  6.25     341
#> 342 -7  8  6.25     342
#> 343 -5  8  6.25     343
#> 344 -3  8  6.25     344
#> 345 -1  8  6.25     345
#> 346  1  8  6.25     346
#> 347  3  8  6.25     347
#> 348  5  8  6.25     348
#> 349  7  8  6.25     349
#> 350  9  8  6.25     350
#> 351 -9 -8  8.75     351
#> 352 -7 -8  8.75     352
#> 353 -5 -8  8.75     353
#> 354 -3 -8  8.75     354
#> 355 -1 -8  8.75     355
#> 356  1 -8  8.75     356
#> 357  3 -8  8.75     357
#> 358  5 -8  8.75     358
#> 359  7 -8  8.75     359
#> 360  9 -8  8.75     360
#> 361 -9 -4  8.75     361
#> 362 -7 -4  8.75     362
#> 363 -5 -4  8.75     363
#> 364 -3 -4  8.75     364
#> 365 -1 -4  8.75     365
#> 366  1 -4  8.75     366
#> 367  3 -4  8.75     367
#> 368  5 -4  8.75     368
#> 369  7 -4  8.75     369
#> 370  9 -4  8.75     370
#> 371 -9  0  8.75     371
#> 372 -7  0  8.75     372
#> 373 -5  0  8.75     373
#> 374 -3  0  8.75     374
#> 375 -1  0  8.75     375
#> 376  1  0  8.75     376
#> 377  3  0  8.75     377
#> 378  5  0  8.75     378
#> 379  7  0  8.75     379
#> 380  9  0  8.75     380
#> 381 -9  4  8.75     381
#> 382 -7  4  8.75     382
#> 383 -5  4  8.75     383
#> 384 -3  4  8.75     384
#> 385 -1  4  8.75     385
#> 386  1  4  8.75     386
#> 387  3  4  8.75     387
#> 388  5  4  8.75     388
#> 389  7  4  8.75     389
#> 390  9  4  8.75     390
#> 391 -9  8  8.75     391
#> 392 -7  8  8.75     392
#> 393 -5  8  8.75     393
#> 394 -3  8  8.75     394
#> 395 -1  8  8.75     395
#> 396  1  8  8.75     396
#> 397  3  8  8.75     397
#> 398  5  8  8.75     398
#> 399  7  8  8.75     399
#> 400  9  8  8.75     400