Converts an Initial Design into a Cell Mapping Grid

This function takes an initial design -- with rows being the observations and columns standing for the dimensions (plus the corresponding objective) -- and adds an additional column to the data.frame. This additional column states the cell ID for each observation.

convertInitDesignToGrid(init, lower, upper, blocks)

Arguments

init
[data.frame] The initial design, consisting of d + 1 columns (d dimensions and one column for the objective value) and one row per observation.

lower
[numeric or integer] The lower limits per dimension.

upper
[numeric or integer] The upper limits per dimension.

blocks
[integer] The number of blocks per dimension.

Value

[data.frame]. A data.frame, which includes an additional column (cell.ID) compared to the initial design (init). The cell.ID will be a value between 1 and prod(blocks).

Examples

# (1) create an initial design:
X = t(replicate(n = 200, expr = runif(n = 5, min = -10, max = 10)))
f = function(x) sum(x^2)
y = apply(X = X, MARGIN = 1, FUN = f)
init = data.frame(X, y = y)

# (2) compute the cell mapping grid
convertInitDesignToGrid(init = init, lower = -10, upper = 10, blocks = 20)

#>               X1          X2          X3          X4          X5         y
#> 1   -8.384997249  6.68666075  2.01521772 -6.85583117 -9.85201118 263.14526
#> 2   -0.672130054 -0.04445223 -4.20465511  4.65763974  5.45043022  69.53366
#> 3    7.492013215 -6.50118746 -9.31517335 -3.59228538 -1.95343523 201.88858
#> 4   -6.086603305 -1.92923765 -8.72677085 -2.22597374  9.51095670 212.33848
#> 5   -4.202154092  3.56760855  4.70639198 -6.08086534  9.61079349 181.88033
#> 6    4.830430583 -8.97107447  0.60424927  3.91647758  3.77112007 133.73850
#> 7   -9.375393493 -5.48874931 -3.98338388  2.72931230 -0.41950901 141.51685
#> 8   -1.356574837  4.12867676  8.97153153 -6.39322464 -5.66200247 172.30624
#> 9    3.603258352 -0.02308779  2.83358697  3.20568698 -8.07951684  96.56824
#> 10   5.312003279  5.39349609  9.81424625  9.41041806 -2.21634479 247.09476
#> 11  -0.776270707 -3.69516495 -6.50648212  0.63147082 -0.12725968  57.00610
#> 12   5.586172519 -5.91643314  4.26794558 -8.69567777 -2.91586402 168.54194
#> 13   6.503988421 -4.52363510  1.40089902 -3.28561839  1.92525578  79.22956
#> 14  -6.169639369  8.95527876  0.84960817  0.89206787 -4.42805692 139.38677
#> 15  -1.065950617 -2.56977631 -9.43878051 -0.68025618 -2.19937225 102.13057
#> 16  -9.598695645 -2.46058145  1.19825680  7.14167172 -2.30380578 155.93623
#> 17   0.558340712  2.01275047 -4.77257283 -4.19899677 -0.39849652  44.93073
#> 18   8.400110924 -1.98559631 -5.73654578  3.43533631 -8.82771778 197.14255
#> 19   9.941382706 -7.01929065  0.37113271  6.92240109  4.36539448 215.21558
#> 20  -5.173719600  0.94086737  6.69603630 -9.44087949 -0.61231400 161.99464
#> 21   6.113600614  6.28102620 -1.92177996 -5.63137983 -1.63277195 114.89902
#> 22   3.377414956  0.15300564  3.20718613  0.23582628  6.71104874  66.81017
#> 23   4.175623227  7.48411882 -9.77040924  7.76499136  9.92693838 327.74796
#> 24   0.003830018 -2.82065951  5.49826045  1.68950502  2.67952742  48.22130
#> 25   7.173323096  1.33788677 -4.94005961  8.37606429  7.34700410 201.78762
#> 26  -5.029226062 -1.94237576  5.39260352 -7.61029251 -6.10610077 153.34713
#> 27  -6.708615036  3.26413163  7.13150009  8.53092897  1.04755189 180.39248
#> 28   1.541313888  3.74895491 -5.10563541 -9.10765683  8.19709113 192.63954
#> 29  -8.586375616  9.93782947  2.23704835 -6.54882309  8.18881930 287.43453
#> 30  -9.250976662  1.87107581 -5.26044890  8.12594533  6.37745968 223.45680
#> 31   3.996587144 -5.59999342  4.55981877 -5.65831075 -0.87539603 100.90738
#> 32  -3.344004834  1.36705337 -4.95588550 -0.71972867  8.35321014 107.90613
#> 33   9.456884358  6.38164893  8.05847596  1.62732088  5.46016962 227.55877
#> 34   9.902460529  4.21942499 -5.70114809 -4.16484740  4.43519459 185.38227
#> 35   7.332314067 -5.23093787 -9.91007384  8.87032928 -1.23725599 259.54865
#> 36   5.012066565  3.35631529 -1.84053597 -2.97502370  4.76183123  71.29904
#> 37   3.285709857 -8.29550600  7.12264313 -8.46033357  7.05689607 251.72038
#> 38  -7.873060782 -0.30394353 -5.05561779  3.73138421 -6.72753604 146.81971
#> 39   9.056495996 -3.56290898 -2.76931766  7.75446835  6.56028836 205.55272
#> 40  -7.986870846  8.12103156  5.45460729 -2.33258659  9.99304914 264.79599
#> 41  -3.014019029  8.94636532 -5.67800048 -9.35814582 -7.09368313 259.25669
#> 42   7.087677759 -5.73701378 -5.79378524 -9.20958620  8.89549606 280.66278
#> 43  -5.101440251  5.62245136 -4.23525664  7.50715821 -4.08499812 148.61868
#> 44   9.670508159  1.79675113  5.18316760  6.72150612  5.25638939 196.42054
#> 45  -1.654601474 -7.23850318 -8.38310077  3.11965252  2.04007718 139.30416
#> 46   3.139916593 -3.41365673  9.58948441  4.30372268  7.45260610 187.53371
#> 47   9.665674940 -5.62874014  3.29060128 -2.20871924 -9.07872722 223.23777
#> 48   2.338291164  1.96949989 -1.86292740  7.16656299  0.35362370  64.30171
#> 49   9.585868209 -9.65968625  3.46895652 -2.57460236  8.36021274 273.75380
#> 50   3.559561893  3.30304923  5.12082176  0.85674310 -5.21423803  77.72572
#> 51   0.177871445 -1.65471257  4.53897706  2.75371104 -2.07180090  35.24731
#> 52   9.189652149 -4.02683940 -8.99597669  1.52374841 -5.64188386 215.74540
#> 53  -7.482874584  8.76305374  6.02550254  5.16107239  0.65130328 196.15207
#> 54   0.936095496 -8.08146992 -2.23300495 -6.55296223  3.81451696 128.66460
#> 55   3.504170072  8.92589695 -6.07560956  9.37275005 -2.25807445 221.81122
#> 56   3.006877974  6.29192407 -8.58070455  0.53660648  5.26949652 150.31365
#> 57  -1.292267153  1.04944672 -5.91938708 -9.37947964  9.39414118 214.03496
#> 58  -6.427738122  5.56585572  7.71421597  6.72892501  2.10736884 181.52313
#> 59   8.137589172 -9.28180382 -7.37162980 -8.11939266  3.93167316 288.09576
#> 60  -1.885425560 -8.68726721 -7.47014756  8.74660436 -5.67239539 243.50570
#> 61   3.321840232 -5.92198639  0.97182878  6.55002829 -7.44884698 145.43719
#> 62  -4.776277645 -4.23165945 -9.70461968  7.07447454 -1.71904343 187.90271
#> 63   5.020276913  4.38027916  0.30710603  9.49672515 -2.78444877 142.42528
#> 64  -6.592089860  5.66281340 -9.38152208  5.78840717 -6.66792645 241.50296
#> 65  -9.425748694  5.71902328  6.50615727  9.30480836 -2.42717642 256.35269
#> 66  -6.512273448  2.07500805  6.10927467 -9.26641182  4.65821333 191.60394
#> 67  -5.694546835 -9.67993712 -7.42794843  3.72126208  2.83882529 203.21018
#> 68  -3.453229596 -2.25036038  4.15000791  3.95310855  8.51313532 122.31202
#> 69  -0.819645510  1.91908404 -6.69978021  1.33819569  7.96154742 114.41876
#> 70   1.889445623  6.63379838  1.86816730  5.57941267 -2.04456706  86.37744
#> 71   6.997655481  4.83691163 -3.64419561 -7.76639552 -7.97809269 209.60992
#> 72   6.001954963 -2.40106516 -8.94598057  9.73017395  2.08347934 220.83632
#> 73  -7.024345542  0.77520344 -7.48421546  9.26363507 -9.06129971 273.87794
#> 74  -6.771816351  8.72974852  9.41362725  4.20548105  7.70181861 287.68646
#> 75   9.537506332 -9.30200713 -1.22485462  2.88939521  9.89664809 285.28388
#> 76  -3.706747587  7.11763932  0.80675363  7.46834705  0.31171239 120.92499
#> 77   7.389185959  7.13045189 -3.08633632 -9.99792617 -5.89640509 249.69501
#> 78   8.904409814 -4.37332974  7.61868434  9.22232553 -9.91963092 339.90924
#> 79   1.058004112 -5.61528055  2.93454287  4.55404093  7.45174885 117.53014
#> 80  -2.360369079  7.85491408  6.87740571  4.59522417 -9.23170047 220.91011
#> 81  -3.323533651  4.98545304  8.26316021 -5.91341874 -0.56488777 139.46805
#> 82  -2.296337993 -2.14496142 -5.63323139 -9.85582119 -2.72575714 146.17429
#> 83   0.522134597  6.88326787 -7.28017658 -2.47065338 -4.73382703 129.16622
#> 84  -0.042940606 -9.60949499 -8.25632380  9.91925329  8.84264411 337.09506
#> 85  -7.512530978 -7.69621396 -2.66623793 -0.57919929  5.68344891 155.41572
#> 86  -0.192143745 -7.96098090  8.86919449  4.12279005  0.04528473 159.07620
#> 87   0.598695530 -4.83147303  4.23506834  9.36925100  3.22155579 139.79866
#> 88   5.925011807 -8.53581736  7.69485769  6.53400754  4.51527919 230.25778
#> 89  -2.653419566  0.46856833 -0.60242175 -8.02867229  0.32069820  72.18553
#> 90  -1.005323450  0.98923096  2.99012380 -6.81930348 -4.12498416  74.44849
#> 91   7.825697339  1.76157252  2.66463127 -4.80981047 -3.63574546 107.79786
#> 92  -3.214922380  5.21522277 -9.78255072 -8.73725322 -4.72316416 231.88045
#> 93   1.876731771 -4.04879942 -8.10754456  5.22510720  5.44953695 142.64638
#> 94  -1.960192136 -8.81834094  5.00495769  0.56644711 -4.48389765 127.08129
#> 95  -4.775248999  0.87662338 -6.01998441  2.90109545  5.00479692  93.27603
#> 96  -4.580443311  7.51243819  7.03486348 -9.08301379 -2.37886282 215.06662
#> 97  -6.333758915  8.43599283  8.12373698 -6.14113364  7.28750553 268.09884
#> 98  -9.734547962  7.71320679 -2.16915786 -8.47647374  2.93562610 239.42874
#> 99   8.333077938  6.69017240  0.02494141  2.04289540 -0.62081932 118.75806
#> 100  7.727484764  3.41096985 -2.12444561  1.02581345 -5.15014648 103.43831
#> 101 -5.935823419 -8.63316593 -3.85452526  9.86236189 -7.67381646 280.77656
#> 102  4.308461533 -5.47782537 -7.33302905  8.54563523  7.88245314 237.50368
#> 103 -5.926712155 -4.86034162  2.28011207 -1.17001625 -3.68810629  78.91881
#> 104 -7.973724781 -4.54707609  3.07398794  8.55740204 -9.46750660 256.56840
#> 105  1.189823942  7.08347304 -8.48233901 -5.88342288 -0.08220707 158.16277
#> 106 -4.762861119 -1.68077099 -1.21700001 -6.35244161 -8.43343468 138.46726
#> 107 -9.644600879  4.56359947  9.41690050  0.70435176 -6.67018464 247.51026
#> 108 -8.836486987 -0.81721962  2.45597643  1.89638403  4.86036475 112.00259
#> 109  0.485602687 -0.06349499 -9.70562395  9.15838249 -0.93821530 179.19520
#> 110  9.630636983  0.89410332 -2.79915198  0.06292209 -8.50732735 173.76242
#> 111  9.395837807  2.97002628  9.69130933  0.72266514  5.22118810 218.80735
#> 112  1.287714541  5.56908607  2.78121977 -2.42078529 -2.80938659  54.16097
#> 113  6.884628325 -0.86501638 -6.41931419  9.32418392  4.26823903 194.51223
#> 114 -9.502279377  2.52912988  6.55842644  2.86138063  4.16968524 165.27654
#> 115  7.414543070 -4.36374837 -6.53550316 -7.80659962  1.72444383 180.64725
#> 116  7.014426547  8.43052353  3.51323804 -2.98823675 -9.82683850 238.11506
#> 117 -1.382024148  4.49870194 -1.75487226  1.01724790 -6.87278860  73.49790
#> 118 -9.139974136  3.69288226  0.28804994  8.91375697  6.86193884 223.80075
#> 119 -8.702338133  5.24221829 -8.43231495 -7.32984675  6.19392402 266.40683
#> 120  1.187664787  4.67459372 -3.00609048  9.65337333 -1.82292240 128.80962
#> 121 -2.782526128  3.05020079 -7.12881548  3.61597231  7.18571787 132.57598
#> 122  6.772599793  7.62625850 -5.24604525  1.57261692  7.95805011 197.35260
#> 123 -3.111419389  5.21716460  4.62207290  7.04864753  9.51636082 198.50785
#> 124 -7.739049983  9.40524559  2.96165564 -8.87853913 -0.53725808 236.24005
#> 125 -4.107499737  2.20599244 -7.57782239  2.58833016  4.23910606 103.83082
#> 126  2.241655733 -9.31141534  3.31781574  3.05934553 -0.96283121 113.02202
#> 127  0.336082038  3.53221594  7.61866188  4.62458309 -2.02760370  96.13145
#> 128 -1.389586926 -7.09185306 -0.97759341  1.57562983 -8.59165759 129.48021
#> 129  4.846523604  0.98441413  2.69088544 -7.26742832  8.60466993 158.55459
#> 130  3.022372792  0.57418545 -5.49769223  9.66533739 -7.81902429 194.24493
#> 131 -8.601089655  5.55291014  3.16644464 -4.81594513  8.36675690 208.03587
#> 132 -7.025793064  2.81360970 -5.97523152  4.87242630  6.95274857 165.06281
#> 133  7.072717459  6.06825145 -9.52246470  5.90755430  7.73101021 272.19206
#> 134  2.000746769 -1.34423667  8.40551606  2.32028856 -1.28563683  83.49926
#> 135  7.946088202  5.38027044 -7.43768483  0.77266241 -7.41967548 203.05537
#> 136  0.219297679  0.87585382 -5.16447820  6.06288231  8.95827449 144.49627
#> 137 -1.852773093  0.29489126  9.63083701 -2.91626544 -0.88319704 105.55739
#> 138 -5.599505873 -8.36333953 -2.61527863 -7.46666832  4.79745011 186.90626
#> 139  9.117325717  3.37349005 -0.63821302  0.76384893 -3.38092596 106.92750
#> 140  2.573642158 -0.94895057  7.76030355  5.68403639 -7.23880898 152.45508
#> 141 -4.373738808  4.13892996 -1.15626160  3.55397046  4.01429971  66.34258
#> 142 -8.631871613  9.17007041 -5.11032958  8.52566669  0.05494154 257.40488
#> 143  8.710342841 -1.22400735 -2.99026927 -0.46991823  9.32319050 173.45268
#> 144 -4.056761460 -4.18703383  7.19444495 -6.13374666  6.18216743 161.59065
#> 145 -2.690269751 -4.94600694  7.12162028 -7.39951261  5.83815321 171.25483
#> 146 -2.881969796  9.28687675 -7.62633582 -0.83678701  0.47300873 153.63678
#> 147  6.741234516 -9.32991925 -9.13343121  5.84205889 -5.94602070 285.39602
#> 148 -8.624091037  7.28882147 -8.45820281 -1.37993163  9.47245116 290.67460
#> 149  1.419146876  5.08813291 -5.15091341  2.54315732 -6.63170794 104.88218
#> 150 -1.104207342 -6.70649532 -7.63446434 -2.65071219  8.93117810 191.27362
#> 151  7.508215266 -4.79882175  3.14596875 -7.56270749 -3.62303922 159.62006
#> 152  2.461220603 -1.06005140  4.16586892  7.80058000 -7.91386992 148.01417
#> 153  2.608234379  6.83117821 -2.71475137  2.45845556 -0.78504972  67.49806
#> 154 -6.731754411  0.27733724 -7.13467123  3.00597402 -6.89410938 152.86159
#> 155  0.166197242 -8.28259111  0.06053326 -9.22856234  6.48065754 195.79789
#> 156 -4.280288517 -5.06498266 -2.04012812  2.64185750  8.97157690 135.60565
#> 157 -3.197615123  2.54731971  6.49830790 -7.09769246  9.34278261 196.60641
#> 158 -8.279220206 -5.19077531  2.77767466  0.06790163 -4.98953481 128.10518
#> 159  9.580675489  3.13159164 -8.20302197  2.61085861  7.53535323 232.48391
#> 160  8.605630714 -1.61028295 -0.08880260 -2.94590616 -5.02530266 110.58981
#> 161  9.174557975 -2.80079755 -6.02027516  4.52570635 -6.63735099 192.79714
#> 162 -5.470373407  6.91080110 -3.08292798 -8.65360343 -9.45961092 251.55769
#> 163 -7.639108608 -9.72609264  4.71458737  8.47921532  9.64381198 340.08039
#> 164 -1.894595921 -4.60610733 -0.23422698 -8.97875095 -9.95046340 204.49027
#> 165 -0.083306502 -5.12417433 -4.88501482  1.04637554 -4.96844743  75.90784
#> 166  2.569003333 -3.67768987  8.49040784  8.12983400 -8.16060409 224.90187
#> 167  4.755903045  1.39481080  3.92977151  4.01535007 -9.20118690 140.79209
#> 168 -1.369551923  6.28462731  1.33172554  1.76494346  7.60618064 104.11472
#> 169  4.386359267  8.72857787 -1.41136772 -4.27625349  5.41286367 145.00562
#> 170  2.329972731 -6.81450505 -1.96803988 -9.96707023 -1.64067929 157.77375
#> 171  5.278687207  5.88955333 -5.14825959  5.18341908 -9.41791693 204.62095
#> 172 -1.253184341  0.60818975 -1.40337298  6.18593918  8.45129237 113.60001
#> 173  3.500469471 -8.12063615 -6.05073366 -0.48909749  4.28743268 133.43069
#> 174  0.761585687  7.50766021 -4.12829174 -5.50469145  3.81835662 118.86924
#> 175 -3.753411905 -0.33403643 -7.56406720  3.80014470 -3.68537191  99.43786
#> 176  6.323753651  0.91830877  8.94634367 -9.22947187  2.95287753 214.77285
#> 177 -8.407619782 -1.34138461  5.21786268 -4.12482320 -7.85336214 178.40294
#> 178 -8.540111165 -2.87370845  1.43121685  3.55976118  3.81438204 110.46149
#> 179 -2.531783576 -8.04043891 -8.01980584  5.49603167 -6.97008883 214.16437
#> 180 -2.107868404  8.69996807  4.79179128  2.71336245 -3.23560299 120.92528
#> 181  9.248326090 -7.76420180  3.00970532  9.67023103 -6.41744991 289.56972
#> 182 -7.240591180 -8.86882828  0.92735238 -0.66990398  7.92756749 195.23736
#> 183 -9.629371334  7.01948458  0.02846846 -6.67676447 -2.77987798 194.30567
#> 184  3.040003679  7.48353671  4.56611320 -1.75054289 -6.81541438 135.60861
#> 185  6.914071091  7.68738204 -2.08003773 -9.23718423  9.02791934 278.05568
#> 186  5.674550300 -5.61120518 -3.38136026 -6.44481342 -2.74967388 124.21607
#> 187 -2.170935166  8.07754701 -4.47086302 -7.21281365 -0.97327391 142.92028
#> 188 -7.793403021  7.61913960  0.67685409  4.78501590  4.39045570 161.41903
#> 189 -4.710361538  3.47712734 -5.80756830  3.80810332 -6.10480969 119.77612
#> 190  5.643082568 -9.69935443  8.68345047 -8.31127734 -9.31895454 357.24441
#> 191 -3.780810125  5.69419448 -4.39198907 -6.69432411 -1.32810373 112.58578
#> 192  6.280903993  3.73027960  5.54612326 -8.17598050  4.85247791 174.51742
#> 193 -2.066520485 -4.07943555  4.85848000 -0.66981121  5.46055383  74.78342
#> 194 -9.557658988  6.30046399 -3.92333158  8.63591730  0.80862029 221.67016
#> 195 -4.903632984 -0.44825273  4.08927983 -9.84988893  5.44817555 167.67169
#> 196 -4.684128715  9.42081872 -6.18949760 -9.60608222  2.24753959 246.33102
#> 197  0.514628179  2.69040085 -2.09104082 -6.50263559 -3.45373677  66.08812
#> 198  2.295063636 -8.41667178  0.38121941 -3.87437381  3.42790609 103.01432
#> 199 -2.126565194  4.95696493 -7.22146957 -2.45313909  2.82766871  95.25701
#> 200  9.883893486 -2.62881884 -3.25302898  8.27370886  8.18840722 250.68851
#>     cell.ID
#> 1     29122
#> 2   2514190
#> 3   1328078
#> 4   3096564
#> 5   3069866
#> 6   2188035
#> 7   1538481
#> 8    671489
#> 9    268994
#> 10  1279916
#> 11  1521330
#> 12  1133696
#> 13  1812517
#> 14   884364
#> 15  1192149
#> 16  1260541
#> 17  1482251
#> 18   265779
#> 19  2372060
#> 20  1446605
#> 21  1315537
#> 22  2645414
#> 23  3176355
#> 24  2014151
#> 25  2866238
#> 26   502165
#> 27  1911064
#> 28  2881872
#> 29  2909182
#> 30  2705821
#> 31  1477694
#> 32  2954227
#> 33  2495540
#> 34  2281900
#> 35  1424098
#> 36  2299476
#> 37  2734834
#> 38   585783
#> 39  2698940
#> 40  3102363
#> 41   321967
#> 42  2881698
#> 43   938305
#> 44  2534240
#> 45  2024449
#> 46  2839734
#> 47    61300
#> 48  1739433
#> 49  2941220
#> 50   726274
#> 51  1221771
#> 52   728520
#> 53  1726763
#> 54  2106831
#> 55  1273574
#> 56  2480734
#> 57  3041829
#> 58  2055104
#> 59  2088819
#> 60   784829
#> 61   452094
#> 62  1416106
#> 63  1276296
#> 64   600304
#> 65  1278701
#> 66  2246644
#> 67  2024805
#> 68  2989747
#> 69  2809430
#> 70  1244732
#> 71   338697
#> 72  2072557
#> 73   153003
#> 74  2839964
#> 75  3139220
#> 76  1740347
#> 77   642758
#> 78   158919
#> 79  2836892
#> 80   118748
#> 81  1479487
#> 82  1121748
#> 83   857131
#> 84  3032410
#> 85  2474843
#> 86  1719250
#> 87  2237711
#> 88  2374836
#> 89  1611808
#> 90   829009
#> 91  1005038
#> 92   808307
#> 93  2520512
#> 94   886029
#> 95  2497406
#> 96  1127146
#> 97  2751564
#> 98  1931141
#> 99  1540339
#> 100  731078
#> 101  474425
#> 102 2864895
#> 103 1028905
#> 104  149303
#> 105 1472752
#> 106  187366
#> 107  567881
#> 108 2332982
#> 109 1592191
#> 110  243020
#> 111 2487860
#> 112 1181112
#> 113 2393397
#> 114 2342641
#> 115 1777318
#> 116   61578
#> 117  571489
#> 118 2708261
#> 119 2576702
#> 120 1434692
#> 121 2825068
#> 122 2809957
#> 123 3181907
#> 124 1453183
#> 125 2337046
#> 126 1549213
#> 127 1239071
#> 128  251649
#> 129 2901015
#> 130  473814
#> 131 2925502
#> 132 2673843
#> 133 2840338
#> 134 1383373
#> 135  401118
#> 136 3009811
#> 137 1503809
#> 138 2258825
#> 139 1043880
#> 140  446993
#> 141 2347486
#> 142 1745982
#> 143 3114979
#> 144 2590906
#> 145 2422908
#> 146 1673188
#> 147  760017
#> 148 3104742
#> 149  577912
#> 150 2936869
#> 151  981318
#> 152  461773
#> 153 1539133
#> 154  585004
#> 155 2564031
#> 156 2978886
#> 157 3062647
#> 158  884882
#> 159 2816680
#> 160  699779
#> 161  593360
#> 162   10725
#> 163 3189603
#> 164   11709
#> 165  890090
#> 166  311333
#> 167  117435
#> 168 2812729
#> 169 2443575
#> 170 1283273
#> 171  121916
#> 172 3011409
#> 173 2313234
#> 174 2114351
#> 175 1064987
#> 176 1927417
#> 177  366162
#> 178 2188542
#> 179  600428
#> 180 1061968
#> 181  637260
#> 182 2796023
#> 183 1148341
#> 184  549954
#> 185 3043157
#> 186 1146496
#> 187 1458368
#> 188 2356343
#> 189  585866
#> 190   15216
#> 191 1306307
#> 192 2254277
#> 193 2477708
#> 194 1746721
#> 195 2405786
#> 196 1921586
#> 197  987051
#> 198 2132033
#> 199 1977088
#> 200 3026560